refactor: clean up firebase init
