--- src/firebase/useFirebase.js
@@ -1,12 +1,0 @@
-import { useEffect, useState } from 'react'
-import getFirebase from './firebase'
-
-export default function useFirebase() {
-  const [instance, setInstance] = useState(null)
-
-  useEffect(() => {
-    setInstance(getFirebase())
-  }, [])
-
-  return instance
-}
